🌱 test: split oversized hooks test files (batch B)
