+++ web/src/hooks/__tests__/useCachedData.security-gitops.fetch.test.ts
@@ -1,0 +1,325 @@
+/**
+ * Deep branch-coverage tests for useCachedData.ts
+ *
+ * Tests the internal utility functions (fetchAPI, fetchClusters,
+ * fetchFromAllClusters, fetchViaSSE, etc.) and every exported
+ * useCached* hook by mocking the underlying cache layer and network.
+ */
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
+
+// ---------------------------------------------------------------------------
+// Mocks — must be declared BEFORE importing the module under test
+// ---------------------------------------------------------------------------
+
+const { mockClusterCacheRef, mockIsDemoMode } = vi.hoisted(() => ({
+  mockClusterCacheRef: { clusters: [] as Array<{ name: string; context?: string; reachable?: boolean; namespaces?: string[] }> },
+  mockIsDemoMode: vi.fn(() => false),
+}))
+
+const mockUseCache = vi.fn()
+const mockIsBackendUnavailable = vi.fn(() => false)
+const mockAuthFetch = vi.fn()
+const mockIsAgentUnavailable = vi.fn(() => true)
+const mockFetchSSE = vi.fn()
+const mockKubectlProxy = {
+  getEvents: vi.fn(),
+  getPodIssues: vi.fn(),
+  exec: vi.fn(),
+}
+const mockSettledWithConcurrency = vi.fn()
+const mockFetchProwJobs = vi.fn()
+const mockFetchLLMdServers = vi.fn()
+const mockFetchLLMdModels = vi.fn()
+
+vi.mock('../../lib/cache', () => ({
+    createCachedHook: vi.fn(),
+  useCache: (...args: unknown[]) => mockUseCache(...args),
+  createCachedHook: (_config: unknown) => () => mockUseCache(_config),
+  REFRESH_RATES: {
+    realtime: 15_000, pods: 30_000, clusters: 60_000,
+    deployments: 60_000, services: 60_000, metrics: 45_000,
+    gpu: 45_000, helm: 120_000, gitops: 120_000,
+    namespaces: 180_000, rbac: 300_000, operators: 300_000,
+    costs: 600_000, default: 120_000,
+  },
+}))
+
+vi.mock('../../lib/api', () => ({
+    createCachedHook: vi.fn(),
+  isBackendUnavailable: () => mockIsBackendUnavailable(),
+  authFetch: (...args: unknown[]) => mockAuthFetch(...args),
+}))
+
+vi.mock('../../lib/demoMode', () => ({
+  isDemoMode: () => mockIsDemoMode(),
+}))
+
+vi.mock('../../lib/kubectlProxy', () => ({
+    createCachedHook: vi.fn(),
+  kubectlProxy: mockKubectlProxy,
+}))
+
+vi.mock('../../lib/sseClient', () => ({
+    createCachedHook: vi.fn(),
+  fetchSSE: (...args: unknown[]) => mockFetchSSE(...args),
+}))
+
+vi.mock('../mcp/shared', () => ({
+    createCachedHook: vi.fn(),
+  clusterCacheRef: mockClusterCacheRef,
+  agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
+}))
+
+vi.mock('../useLocalAgent', () => ({
+    createCachedHook: vi.fn(),
+  isAgentUnavailable: () => mockIsAgentUnavailable(),
+}))
+
+vi.mock('../../lib/constants', async (importOriginal) => {
+  const actual = await importOriginal() as Record<string, unknown>
+  return { ...actual,
+  LOCAL_AGENT_HTTP_URL: 'http://localhost:8765',
+  STORAGE_KEY_TOKEN: 'kc_token',
+} })
+
+vi.mock('../../lib/constants/network', async (importOriginal) => {
+  const actual = await importOriginal() as Record<string, unknown>
+  return { ...actual,
+  FETCH_DEFAULT_TIMEOUT_MS: 10_000,
+  AI_PREDICTION_TIMEOUT_MS: 30_000,
+  KUBECTL_EXTENDED_TIMEOUT_MS: 60_000,
+} })
+
+vi.mock('../../lib/utils/concurrency', () => ({
+    createCachedHook: vi.fn(),
+  settledWithConcurrency: async (...args: unknown[]) => {
+    const result = await mockSettledWithConcurrency(...args)
+    // Invoke the onSettled callback (3rd arg) so the production code's
+    // accumulation logic runs.  Without this, tests that use mockResolvedValue
+    // silently skip the callback and return empty results.
+    const onSettled = args[2] as ((r: PromiseSettledResult<unknown>, i: number) => void) | undefined
+    if (onSettled && Array.isArray(result)) {
+      result.forEach((r: PromiseSettledResult<unknown>, i: number) => onSettled(r, i))
+    }
+    return result
+  },
+}))
+
+vi.mock('../useCachedProw', () => ({
+    createCachedHook: vi.fn(),
+  fetchProwJobs: (...args: unknown[]) => mockFetchProwJobs(...args),
+}))
+
+vi.mock('../useCachedLLMd', () => ({
+    createCachedHook: vi.fn(),
+  fetchLLMdServers: (...args: unknown[]) => mockFetchLLMdServers(...args),
+  fetchLLMdModels: (...args: unknown[]) => mockFetchLLMdModels(...args),
+}))
+
+vi.mock('../useCachedISO27001', () => ({
+    createCachedHook: vi.fn(),}))
+
+// Stub the re-exports so the module loads cleanly
+vi.mock('../useWorkloads', () => ({
+    createCachedHook: vi.fn(),}))
+
+vi.mock('../../lib/schemas/validate', () => ({
+    createCachedHook: vi.fn(),
+  validateResponse: (_schema: unknown, data: unknown) => data,
+  validateArrayResponse: (_schema: unknown, data: unknown) => data,
+}))
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+/** Default shape returned by our mocked useCache */
+function makeCacheResult<T>(data: T, overrides?: Record<string, unknown>) {
+  return {
+    data,
+    isLoading: false,
+    isRefreshing: false,
+    isDemoFallback: false,
+    error: null,
+    isFailed: false,
+    consecutiveFailures: 0,
+    lastRefresh: Date.now(),
+    refetch: vi.fn(),
+    ...overrides,
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Tests
+// ---------------------------------------------------------------------------
+
+describe('useCachedData', () => {
+  let mod: typeof import('../useCachedData')
+
+  beforeEach(() => {
+    vi.resetModules()
+    vi.clearAllMocks()
+    localStorage.clear()
+    // Set a valid token so fetchAPI doesn't throw
+    localStorage.setItem('kc_token', 'test-jwt-token')
+    // Reset the shared cluster cache so tests start with a clean slate
+    mockClusterCacheRef.clusters = []
+    mockIsDemoMode.mockReturnValue(false)
+    // Default useCache implementation
+    mockUseCache.mockImplementation((opts: { initialData: unknown }) =>
+      makeCacheResult(opts.initialData)
+    )
+    // Default settledWithConcurrency: run tasks and return settled results
+    mockSettledWithConcurrency.mockImplementation(async (tasks: Array<() => Promise<unknown>>) => {
+      return Promise.allSettled(tasks.map(t => t()))
+    })
+  })
+
+  afterEach(() => {
+    vi.restoreAllMocks()
+  })
+
+  // Lazy-load module after mocks are set up
+  async function loadModule() {
+    mod = await import('../useCachedData')
+    return mod
+  }
+
+  // ========================================================================
+  // useCachedPods
+  // ========================================================================
+
+  // ========================================================================
+  // Security issues via kubectl scanning
+  // ========================================================================
+  describe('security issues kubectl scanning', () => {
+    it('useCachedSecurityIssues fetcher: agent kubectl finds privileged containers', async () => {
+      let capturedOpts: Record<string, unknown> = {}
+      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
+        capturedOpts = opts
+        return makeCacheResult([])
+      })
+
+      // Mutate the shared mock ref directly — avoids the `vi.doMock` +
+      // `resetModules` race that caused kubestellar/console#9305.
+      mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }]
+      mockIsAgentUnavailable.mockReturnValue(false)
+
+      mockKubectlProxy.exec.mockResolvedValue({
+        exitCode: 0,
+        output: JSON.stringify({
+          items: [
+            {
+              metadata: { name: 'bad-pod', namespace: 'default' },
+              spec: {
+                containers: [
+                  { securityContext: { privileged: true } },
+                ],
+                hostNetwork: true,
+                hostPID: true,
+                hostIPC: true,
+              },
+            },
+          ],
+        }),
+      })
+
+      const { useCachedSecurityIssues } = await loadModule()
+      useCachedSecurityIssues()
+
+      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ issue: string; severity: string }>>
+      const issues = await fetcher()
+
+      const issueTypes = issues.map(i => i.issue)
+      expect(issueTypes).toContain('Privileged container')
+      expect(issueTypes).toContain('Host network enabled')
+      expect(issueTypes).toContain('Host PID enabled')
+      expect(issueTypes).toContain('Host IPC enabled')
+    })
+
+    it('useCachedSecurityIssues fetcher: detects root user and missing security context', async () => {
+      let capturedOpts: Record<string, unknown> = {}
+      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
+        capturedOpts = opts
+        return makeCacheResult([])
+      })
+
+      vi.doMock('../mcp/shared', () => ({
+        clusterCacheRef: {
+          clusters: [{ name: 'prod', context: 'prod-ctx', reachable: true }],
+        },
+        agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
+      }))
+      mockIsAgentUnavailable.mockReturnValue(false)
+
+      mockKubectlProxy.exec.mockResolvedValue({
+        exitCode: 0,
+        output: JSON.stringify({
+          items: [
+            {
+              metadata: { name: 'root-pod', namespace: 'apps' },
+              spec: {
+                securityContext: { runAsUser: 0 },
+                containers: [
+                  { securityContext: { runAsUser: 0 } },
+                ],
+              },
+            },
+          ],
+        }),
+      })
+
+      const { useCachedSecurityIssues } = await loadModule()
+      useCachedSecurityIssues()
+
+      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ issue: string }>>
+      const issues = await fetcher()
+      const issueTypes = issues.map(i => i.issue)
+      expect(issueTypes).toContain('Running as root')
+    })
+
+    it('useCachedSecurityIssues fetcher: detects capabilities not dropped', async () => {
+      let capturedOpts: Record<string, unknown> = {}
+      mockUseCache.mockImplementation((opts: Record<string, unknown>) => {
+        capturedOpts = opts
+        return makeCacheResult([])
+      })
+
+      vi.doMock('../mcp/shared', () => ({
+        clusterCacheRef: {
+          clusters: [{ name: 'prod', context: 'prod-ctx', reachable: true }],
+        },
+        agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
+      }))
+      mockIsAgentUnavailable.mockReturnValue(false)
+
+      mockKubectlProxy.exec.mockResolvedValue({
+        exitCode: 0,
+        output: JSON.stringify({
+          items: [
+            {
+              metadata: { name: 'cap-pod', namespace: 'system' },
+              spec: {
+                containers: [
+                  {
+                    securityContext: {
+                      capabilities: { add: ['NET_ADMIN'], drop: [] },
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+        }),
+      })
+
+      const { useCachedSecurityIssues } = await loadModule()
+      useCachedSecurityIssues()
+
+      const fetcher = capturedOpts.fetcher as () => Promise<Array<{ issue: string }>>
+      const issues = await fetcher()
+      const issueTypes = issues.map(i => i.issue)
+      expect(issueTypes).toContain('Capabilities not dropped')
+    })
+})
+})
