+++ web/src/hooks/__tests__/useUsers.core.k8s.test.ts
@@ -101,450 +101,6 @@ const EMPTY_CLUSTERS: Array<{ name: string }> = []
 // useConsoleUsers
 // =========================================================================
 
-describe('useConsoleUsers', () => {
-  it('fetches users from API on mount and returns them', async () => {
-    const apiUsers = [
-      {
-        id: '1',
-        github_id: '111',
-        github_login: 'alice',
-        email: 'alice@co.com',
-        role: 'admin',
-        onboarded: true,
-        created_at: '2024-01-01T00:00:00Z',
-      },
-      {
-        id: '2',
-        github_id: '222',
-        github_login: 'bob',
-        role: 'viewer',
-        onboarded: false,
-        created_at: '2024-02-01T00:00:00Z',
-      },
-    ]
-    mockGet.mockResolvedValue({ data: apiUsers })
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual(apiUsers)
-    expect(result.current.error).toBeNull()
-    expect(result.current.isRefreshing).toBe(false)
-    expect(mockGet).toHaveBeenCalledWith('/api/users')
-  })
-
-  it('returns demo data when demo mode is on (no API call)', async () => {
-    mockGetDemoMode.mockReturnValue(true)
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-    const logins = result.current.users.map((u) => u.github_login)
-    expect(logins).toContain('admin-user')
-    expect(mockGet).not.toHaveBeenCalled()
-  })
-
-  it('sets error message on API failure and empties users', async () => {
-    mockGet.mockRejectedValue(new Error('Network error'))
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-    expect(result.current.error).toBe('Network error')
-  })
-
-  it('handles non-Error rejection (string message)', async () => {
-    mockGet.mockRejectedValue('server down')
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.error).toBe('Failed to load users')
-  })
-
-  it('handles null data from API gracefully', async () => {
-    mockGet.mockResolvedValue({ data: null })
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-    expect(result.current.error).toBeNull()
-  })
-
-  it('updateUserRole calls PUT and updates local state', async () => {
-    const users = [
-      {
-        id: 'u1',
-        github_id: '111',
-        github_login: 'alice',
-        role: 'viewer' as const,
-        onboarded: true,
-        created_at: '2024-01-01',
-      },
-    ]
-    mockGet.mockResolvedValue({ data: users })
-    mockPut.mockResolvedValue({ data: {} })
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    await act(async () => {
-      const ok = await result.current.updateUserRole('u1', 'admin')
-      expect(ok).toBe(true)
-    })
-
-    expect(mockPut).toHaveBeenCalledWith('/api/users/u1/role', {
-      role: 'admin',
-    })
-    expect(result.current.users[0].role).toBe('admin')
-  })
-
-  it('deleteUser calls DELETE and removes user from local state', async () => {
-    const users = [
-      {
-        id: 'u1',
-        github_id: '1',
-        github_login: 'a',
-        role: 'viewer' as const,
-        onboarded: true,
-        created_at: '2024-01-01',
-      },
-      {
-        id: 'u2',
-        github_id: '2',
-        github_login: 'b',
-        role: 'editor' as const,
-        onboarded: true,
-        created_at: '2024-01-01',
-      },
-    ]
-    mockGet.mockResolvedValue({ data: users })
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.users).toHaveLength(2)
-
-    await act(async () => {
-      const ok = await result.current.deleteUser('u1')
-      expect(ok).toBe(true)
-    })
-
-    expect(mockDelete).toHaveBeenCalledWith('/api/users/u1')
-    expect(result.current.users).toHaveLength(1)
-    expect(result.current.users[0].id).toBe('u2')
-  })
-
-  it('refetch reloads data from the API', async () => {
-    mockGet.mockResolvedValueOnce({
-      data: [
-        {
-          id: '1',
-          github_id: '1',
-          github_login: 'a',
-          role: 'viewer',
-          onboarded: true,
-          created_at: '2024-01-01',
-        },
-      ],
-    })
-
-    const { useConsoleUsers } = await getHooks()
-    const { result } = renderHook(() => useConsoleUsers())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.users).toHaveLength(1)
-
-    mockGet.mockResolvedValueOnce({
-      data: [
-        {
-          id: '1',
-          github_id: '1',
-          github_login: 'a',
-          role: 'viewer',
-          onboarded: true,
-          created_at: '2024-01-01',
-        },
-        {
-          id: '2',
-          github_id: '2',
-          github_login: 'b',
-          role: 'admin',
-          onboarded: true,
-          created_at: '2024-02-01',
-        },
-      ],
-    })
-
-    await act(async () => {
-      await result.current.refetch()
-    })
-
-    expect(result.current.users).toHaveLength(2)
-  })
-})
-
-// =========================================================================
-// useUserManagementSummary
-// =========================================================================
-
-describe('useUserManagementSummary', () => {
-  it('fetches summary from API and returns it', async () => {
-    const summaryData = {
-      consoleUsers: { total: 10, admins: 2, editors: 5, viewers: 3 },
-      k8sServiceAccounts: { total: 20, clusters: ['c1', 'c2'] },
-      currentUserPermissions: [
-        {
-          cluster: 'c1',
-          isClusterAdmin: true,
-          canCreateServiceAccounts: true,
-          canManageRBAC: true,
-          canViewSecrets: true,
-        },
-      ],
-    }
-    mockGet.mockResolvedValue({ data: summaryData })
-
-    const { useUserManagementSummary } = await getHooks()
-    const { result } = renderHook(() => useUserManagementSummary())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.summary).toEqual(summaryData)
-    expect(result.current.error).toBeNull()
-    expect(mockGet).toHaveBeenCalledWith('/api/users/summary')
-  })
-
-  it('returns demo data in demo mode without calling API', async () => {
-    mockGetDemoMode.mockReturnValue(true)
-
-    const { useUserManagementSummary } = await getHooks()
-    const { result } = renderHook(() => useUserManagementSummary())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.summary).not.toBeNull()
-    expect(result.current.summary!.consoleUsers.total).toBe(4)
-    expect(result.current.summary!.consoleUsers.admins).toBe(1)
-    expect(mockGet).not.toHaveBeenCalled()
-  })
-
-  it('falls back to demo data on API error', async () => {
-    mockGet.mockRejectedValue(new Error('Server error'))
-
-    const { useUserManagementSummary } = await getHooks()
-    const { result } = renderHook(() => useUserManagementSummary())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.summary).not.toBeNull()
-    expect(result.current.summary!.consoleUsers.total).toBe(4)
-  })
-
-  it('refetch reloads summary', async () => {
-    const summaryData = {
-      consoleUsers: { total: 5, admins: 1, editors: 2, viewers: 2 },
-      k8sServiceAccounts: { total: 8, clusters: ['c1'] },
-      currentUserPermissions: [],
-    }
-    mockGet.mockResolvedValue({ data: summaryData })
-
-    const { useUserManagementSummary } = await getHooks()
-    const { result } = renderHook(() => useUserManagementSummary())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    const updatedSummary = {
-      ...summaryData,
-      consoleUsers: { ...summaryData.consoleUsers, total: 15 },
-    }
-    mockGet.mockResolvedValue({ data: updatedSummary })
-
-    await act(async () => {
-      await result.current.refetch()
-    })
-
-    expect(result.current.summary!.consoleUsers.total).toBe(15)
-  })
-})
-
-// =========================================================================
-// useOpenShiftUsers
-// =========================================================================
-
-describe('useOpenShiftUsers', () => {
-  it('fetches OpenShift users for a cluster', async () => {
-    const osUsers = [
-      {
-        name: 'admin',
-        fullName: 'Admin',
-        identities: ['htpasswd:admin'],
-        groups: [],
-        cluster: 'prod',
-      },
-      { name: 'dev', cluster: 'prod' },
-    ]
-    mockGet.mockResolvedValue({ data: osUsers })
-
-    const { useOpenShiftUsers } = await getHooks()
-    const { result } = renderHook(() => useOpenShiftUsers('prod'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual(osUsers)
-    expect(mockGet).toHaveBeenCalledWith('/api/openshift/users?cluster=prod')
-  })
-
-  it('returns empty array when no cluster is provided', async () => {
-    const { useOpenShiftUsers } = await getHooks()
-    const { result } = renderHook(() => useOpenShiftUsers(undefined))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-    expect(mockGet).not.toHaveBeenCalled()
-  })
-
-  it('falls back to demo data on API error', async () => {
-    mockGet.mockRejectedValue(new Error('Connection refused'))
-
-    const { useOpenShiftUsers } = await getHooks()
-    const { result } = renderHook(() => useOpenShiftUsers('staging'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users.length).toBeGreaterThan(0)
-    expect(result.current.users[0].cluster).toBe('staging')
-  })
-
-  it('handles null data from API', async () => {
-    mockGet.mockResolvedValue({ data: null })
-
-    const { useOpenShiftUsers } = await getHooks()
-    const { result } = renderHook(() => useOpenShiftUsers('c1'))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-  })
-
-  it('clears users when cluster changes to undefined', async () => {
-    mockGet.mockResolvedValue({
-      data: [{ name: 'admin', cluster: 'c1' }],
-    })
-
-    const { useOpenShiftUsers } = await getHooks()
-    const { result, rerender } = renderHook(
-      ({ cluster }: { cluster?: string }) => useOpenShiftUsers(cluster),
-      { initialProps: { cluster: 'c1' } },
-    )
-
-    await waitFor(() => expect(result.current.users).toHaveLength(1))
-
-    rerender({ cluster: undefined })
-
-    await waitFor(() => expect(result.current.users).toEqual([]))
-  })
-})
-
-// =========================================================================
-// useAllOpenShiftUsers
-// =========================================================================
-
-describe('useAllOpenShiftUsers', () => {
-  it('fetches users from all clusters and aggregates them', async () => {
-    mockGet.mockImplementation((url: string) => {
-      if (url.includes('cluster=c1')) {
-        return Promise.resolve({
-          data: [{ name: 'admin', cluster: 'c1' }],
-        })
-      }
-      if (url.includes('cluster=c2')) {
-        return Promise.resolve({
-          data: [{ name: 'dev', cluster: 'c2' }],
-        })
-      }
-      return Promise.resolve({ data: [] })
-    })
-
-    const { useAllOpenShiftUsers } = await getHooks()
-    const clusters = [{ name: 'c1' }, { name: 'c2' }]
-    const { result } = renderHook(() => useAllOpenShiftUsers(clusters))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toHaveLength(2)
-    expect(result.current.failedClusters).toEqual([])
-  })
-
-  it('returns empty when clusters array is empty', async () => {
-    const { useAllOpenShiftUsers } = await getHooks()
-    // Use stable reference to avoid infinite re-renders
-    const { result } = renderHook(() => useAllOpenShiftUsers(EMPTY_CLUSTERS))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-  })
-
-  it('marks failed clusters and adds demo data for them', async () => {
-    mockGet.mockImplementation((url: string) => {
-      if (url.includes('cluster=good')) {
-        return Promise.resolve({
-          data: [{ name: 'real-user', cluster: 'good' }],
-        })
-      }
-      if (url.includes('cluster=bad')) {
-        return Promise.reject(new Error('unreachable'))
-      }
-      return Promise.resolve({ data: [] })
-    })
-
-    const { useAllOpenShiftUsers } = await getHooks()
-    const clusters = [{ name: 'good' }, { name: 'bad' }]
-    const { result } = renderHook(() => useAllOpenShiftUsers(clusters))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users.length).toBeGreaterThan(1)
-    expect(result.current.failedClusters).toContain('bad')
-  })
-
-  it('handles null data from API for a cluster', async () => {
-    mockGet.mockResolvedValue({ data: null })
-
-    const { useAllOpenShiftUsers } = await getHooks()
-    const clusters = [{ name: 'c1' }]
-    const { result } = renderHook(() => useAllOpenShiftUsers(clusters))
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.users).toEqual([])
-    expect(result.current.failedClusters).toEqual([])
-  })
-})
-
-// =========================================================================
-// useK8sUsers
-// =========================================================================
-
 describe('useK8sUsers', () => {
   it('fetches K8s users for a cluster', async () => {
     const k8sUsers = [
@@ -598,11 +154,6 @@ describe('useK8sUsers', () => {
     expect(result.current.users).toEqual([])
   })
 })
-
-// =========================================================================
-// useK8sServiceAccounts
-// =========================================================================
-
 describe('useK8sServiceAccounts', () => {
   it('fetches service accounts for a cluster', async () => {
     const sas = [
@@ -735,11 +286,6 @@ describe('useK8sServiceAccounts', () => {
     })
   })
 })
-
-// =========================================================================
-// useAllK8sServiceAccounts
-// =========================================================================
-
 describe('useAllK8sServiceAccounts', () => {
   it('fetches service accounts from all clusters', async () => {
     mockGet.mockImplementation((url: string) => {
@@ -812,11 +358,6 @@ describe('useAllK8sServiceAccounts', () => {
     expect(result.current.failedClusters).toEqual([])
   })
 })
-
-// =========================================================================
-// useK8sRoles
-// =========================================================================
-
 describe('useK8sRoles', () => {
   it('fetches roles for a cluster', async () => {
     const roles = [
@@ -888,11 +429,6 @@ describe('useK8sRoles', () => {
     expect(result.current.roles).toEqual([])
   })
 })
-
-// =========================================================================
-// useK8sRoleBindings
-// =========================================================================
-
 describe('useK8sRoleBindings', () => {
   it('fetches bindings for a cluster', async () => {
     const bindings = [
@@ -1026,11 +562,6 @@ describe('useK8sRoleBindings', () => {
     expect(result.current.bindings).toEqual([])
   })
 })
-
-// =========================================================================
-// useClusterPermissions
-// =========================================================================
-
 describe('useClusterPermissions', () => {
   // #7993 Phase 6: useClusterPermissions now calls kc-agent
   // (LOCAL_AGENT_HTTP_URL/rbac/permissions) directly via fetch instead of
